- Export all interfaces from index.ts so that they available from single import - Update `build:clean` script to clean 'dist' folder instead of deleting. It helps avoid rimraf fails when `dist` folder is occupied
