+++ package.json
@@ -14,7 +14,7 @@
     "coverage": "nyc --reporter=lcov --reporter=text npm test",
     "lint": "eslint src/**/*.* tests/*.* --fix",
     "check-types": "tsc",
-    "build:clean": "rimraf dist",
+    "build:clean": "rimraf dist/**/*",
     "build:compile": "babel src --out-dir dist --extensions \".ts\" --ignore src/*.d.ts",
     "build:copy": "copyfiles -u 1 src/vendor/*.* src/tree-viz.ejs dist",
     "build:format": "prettier dist/*.js --write",
